% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreSig-package.R
\docType{package}
\name{PreSig-package}
\alias{PreSig}
\alias{PreSig-package}
\title{PreSig: Computing gene set enrichment and modeling drug response}
\description{
PreSig allow users to compute gene set enrichement scores for single cell gene expression profiles using two existing methods: GSVA and UniPath. It enables cell type assignent to invidual cells based on highest score of cell type specific gene sets. Further it allows model drug response in form of IC50 values using bulk gene expression profiles.
}
\keyword{internal}
